+++ lib/shrine/plugins/restore_cached_data.rb
@@ -1,28 +1,25 @@
 class Shrine
   module Plugins
-    # The `restore_cached_data` plugin re-extracts cached file's metadata on
-    # assignment. This happens when an uploaded file is retained on validation
-    # errors, or when assigning direct uploaded files. In both cases you
-    # usually want to re-extract metadata on the server side, mainly to prevent
+    # The `restore_cached_data` plugin re-extracts metadata when assigning
+    # already cached files, i.e. when the attachment has been retained on
+    # validation errors or assigned from a direct upload. In both cases you may
+    # want to re-extract metadata on the server side, mainly to prevent
     # tempering, but also in case of direct uploads to obtain metadata that
     # couldn't be extracted on the client side.
     #
     #     plugin :restore_cached_data
     #
-    # This will give an opened `UploadedFile` for metadata extraction. For
-    # remote storages this will make an HTTP request, and since metadata is
-    # typically found in the beginning of the file, Shrine will download only
-    # the amount of bytes necessary for extracting the metadata.
+    # It uses the `refresh_metadata` plugin to re-extract metadata.
     module RestoreCachedData
+      def self.load_dependencies(uploader, *)
+        uploader.plugin :refresh_metadata
+      end
+
       module AttacherMethods
         private
 
         def assign_cached(cached_file)
-          uploaded_file(cached_file) do |file|
-            real_metadata = file.open { cache.extract_metadata(file, context) }
-            file.metadata.update(real_metadata)
-          end
-
+          uploaded_file(cached_file) { |file| file.refresh_metadata!(context) }
           super(cached_file)
         end
       end